generate_code/requirements.txt
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
openai>=1.0.0
datasets>=2.14.0
nltk>=3.8
rouge_score>=0.1.2
scikit-learn>=1.3.0
numpy>=1.24.0
pyyaml>=6.0
tqdm
pandas
